requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-groq>=0.1.0
python-dotenv>=1.0.0
requests>=2.31.0
linkedin-api>=2.0.0
schedule>=1.2.0
pydantic>=2.6.0
typing-extensions>=4.9.0
aiohttp>=3.9.0 